requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
astroid==2.11.2
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
cssselect2==0.5.0
dill==0.3.4
Django==4.0.4
factory-boy==3.2.1
Faker==13.11.1
future==0.18.2
html5lib==1.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.22.3
oscrypto==1.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg==3.0.11
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
pylint==2.13.5
pyparsing==3.0.9
PyPDF3==1.0.6
pytest==6.2.5
pytest-base-url==1.4.2
pytest-django==4.5.2
pytest-factoryboy==2.3.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-selenium==3.0.0
pytest-variables==1.9.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tenacity==6.3.1
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
wrapt==1.14.0
xhtml2pdf==0.2.7
